pyproject.toml
[tool.poetry]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = ["Constantine Putilov <putilov.constantine@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
django-rest-framework = "^0.1.0"
envparse = "^0.2.0"
django-filter = "^24.1"
django-debug-toolbar = "^4.3.0"
drf-spectacular = "^0.27.1"
psycopg2 = "^2.9.9"
drf-orjson-renderer = "^1.7.2"


[tool.poetry.group.test.dependencies]
pytest-factoryboy = "^2.7.0"
pytest = "^8.1.1"
pytest-django = "^4.8.0"
pytest-cov = "^4.1.0"
factory-boy = "^3.3.0"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
